refactor(engineer_worktree): split tests.rs (#1266)
tests.rs 732 -> 459 LOC; new tests_extra.rs (~373).
