+++ src/engineer_worktree/mod.rs
@@ -34,6 +34,8 @@ use crate::error::SimardError;
 
 #[cfg(test)]
 mod tests;
+#[cfg(test)]
+mod tests_extra;
 
 /// Subdirectory under the supervisor state root that holds all engineer worktrees.
 pub const WORKTREES_SUBDIR: &str = "engineer-worktrees";
